requirements.txt
aiofiles==0.6.0
certifi==2020.12.5
h11==0.9.0
httpcore==0.11.1
httptools==0.1.1
httpx==0.15.4
idna==3.1
multidict==5.1.0
rfc3986==1.4.0
sanic==20.12.1
sniffio==1.2.0
websockets==8.1
python-dotenv~=0.15.0
SQLAlchemy~=1.3.22
alembic~=1.5.2
marshmallow~=3.10.0
bcrypt~=3.2.0
PyJWT~=2.0.1